BaseTransaction:   * add `_type` and aliases   * move EIP-2930 aliases for senderS, senderR, yParity so they can be available for all transactions
